requirements.txt
frida-tools==6.0.1
androguard==3.3.5
